# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Urbano 2
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /rdg/citas:
    x-swagger-router-controller: rdg
    get:
      summary: Trae mails de una cita a reunion de RDG
      operationId: getCitas
      parameters:
        - in: query
          name: idInstancia
          type: string
          required: false
      tags:
        - RDG - Citas
      responses: 
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Cita'
    post:
      summary: Guarda el mail de una nueva cita de reunion de RDG
      tags: 
        - RDG - Citas
      operationId: createCita
      parameters:
        - in: body
          name: body
          description: Datos de los asistente de un reunion de RDG
          schema:
            type: object
            required:
              - para
              - asunto
              - mensajeHtml
              - idInstancia
              - version
              - fecha
            properties:
              para:
                type: string
              cc:
                type: ['string','null']
              cco:
                type: ['string','null']
              usuario:
                type: ['string','null']
              asunto:
                type: string
              mensajeHtml:
                type: string
              idInstancia:
                type: string
              version:
                type: string
              fecha:
                type: string
      responses:
        200:
          description: ok
  /rdg/citas/listarCorreos:
    x-swagger-router-controller: rdg
    post:
      summary: Trear los correos que estan guardados en serie y maestros
      operationId: getDatosCitas
      parameters:
        - in: body
          name: body
          description: Datos de los asistente de un reunion de RDG
          schema:
            type: object
            required:
              - idInstancia
            properties:
              idInstancia:
                type: string
      tags:
        - RDG - Citas
      responses: 
        200:
          description: OK
          schema:
            $ref: '#/definitions/Cita'
  /rdg/citas/{id}:
    x-swagger-router-controller: rdg
    get:
      summary: Trear una cita de reunion de RDG por ID de la reunion
      operationId: getCitaPorId
      parameters:
        - name: id 
          in: path
          description: ID de la reunion
          type: string
          required: true 
      tags:
        - RDG - Citas
      responses: 
        200:
          description: OK
          schema:
            $ref: '#/definitions/Cita'
            
  # //////////////////////////////////////////////////////////////////// RDG - MInutas
...
  /rdg/minutas/{id}:
    x-swagger-router-controller: rdg
    get:
      summary: Trear una minuta de reunion de RDG por ID
      operationId: getMinutaPorId
      parameters:
        - name: id 
          in: path
          description: ID de la minuta
          type: string
          required: true 
      tags:
        - RDG - Minutas
      responses: 
        200:
          description: OK
          schema:
            $ref: '#/definitions/Minuta'
  # //////////////////////////////////////////////////////////////////// RDG - Serie de Reuniones
  /rdg/series:
    x-swagger-router-controller: rdg
    get:
      summary: Trae serie de reuniones de RDG
      operationId: getSeries
      parameters:
        - in: query
          name: tipo
          type: string
          required: false
        - in: query
          name: jurisdiccion
          type: string
          required: false
      tags:
        - RDG - Series de Reuniones
      responses: 
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SerieReunion'
    post:
      summary: Guarda una nueva serie de reunion de RDG
      tags: 
        - RDG - Series de Reuniones
      operationId: createSerie
      parameters:
        - in: body
          name: body
          description: Datos de la serie de reunion de RDG
          schema:
            type: object
            required:
              - nombre
              - tipo
              - frecuencia
            properties:
              nombre:
                type: string
              tipo:
                type: string
              frecuencia:
                type: string
      responses:
        200:
          description: ok
  # //////////////////////////////////////////////////////////////////// RDG - Serie de Reuniones - Maestro
  /rdg/maestros:
    x-swagger-router-controller: rdg
    get:
      summary: Trae serie de reuniones de RDG
      operationId: getMaestros
      parameters:
        - in: query
          name: tipo
          type: string
          required: false
      tags:
        - RDG - Series Maestros
      responses: 
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/SerieReunion'
  /rdg/maestros/{id}:
    x-swagger-router-controller: rdg
    put:
      summary: Trear una minuta de reunion de RDG por ID
      operationId: editMaestro
      parameters:
        - name: id 
          in: path
          description: ID de la serie Maestro
          type: string
          required: true 
        - in: body
          name: body
          description: Datos para actualizar el proyecto
          schema:
            type: object
            properties:
              nombre: 
                type: string
              tipo:
                type: string
              frecuencia:
                type: string
              participantes:
                type: array
                items:
                  type: object
              cita:
                type: object
                properties:
                  para:
                    type: array
                    items:
                      type: object
                  cc:
                    type: array
                    items:
                      type: object
                  cco:
                    type: array
                    items:
                      type: object
                  exclusivos:
                    type: array
                    items:
                      type: object
              temario:
                type: object
                properties:
                  para:
                    type: array
                    items:
                      type: object
                  cc:
                    type: array
                    items:
                      type: object
                  cco:
                    type: array
                    items:
                      type: object
                  exclusivos:
                    type: array
                    items:
                      type: object
              minuta:
                type: object
                properties:
                    para:
                      type: array
                      items:
                        type: object
                    cc:
                      type: array
                      items:
                        type: object
                    cco:
                      type: array
                      items:
                        type: object
                    exclusivos:
                      type: array
                      items:
                        type: object
              propuestaTemario:
                type: object
                properties:
                  para:
                    type: array
                    items:
                      type: object
                  cc:
                    type: array
                    items:
                      type: object
                  cco:
                    type: array
                    items:
                      type: object
                  exclusivos:
                    type: array
                    items:
                      type: object
      tags:
        - RDG - Series Maestros
      responses: 
        200:
          description: OK
          schema:
            $ref: '#/definitions/Minuta'
  # //////////////////////////////////////////////////////////////////// RDG - Tipos de Reuniones
...
  /rdg/reuniones:
    x-swagger-router-controller: rdg
    get:
      summary: Trae tipos de reuniones de RDG
      operationId: getReuniones
      tags:
        - RDG - Reuniones
      responses: 
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Reunion'
    post:
      summary: Guarda un tipo de reunion de RDG
      tags: 
        - RDG - Reuniones
      operationId: createReunion
      parameters:
        - in: body
          name: body
          description: Datos de la serie de reunion de RDG
          schema:
            type: object
            required:
              - reunion
              - usuarioCreacion
              - fechaCreacion
            properties:
              reunion:
                type: string
              fecha:
                type: string
              desdeHora:
                type: string
              hastaHora:
                type: string
              desdeDate:
                type: integer
              hastaDate:
                type: integer
              usuarioCreacion:
                type: string
              fechaCreacion:
                type: string
      responses:
        200:
          description: ok
  /rdg/reuniones/{id}:
    x-swagger-router-controller: rdg
    get:
      summary: Trae Reunion de RDG por ID
      operationId: getReunionPorId
      parameters:
        - name: id 
          in: path
          description: ID de la Reunion
          type: string
          required: true 
      tags:
        - RDG - Reuniones
      responses: 
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Reunion'
    put:
      summary: Editar una reunion de RDG por ID
      operationId: updateReunion
      tags:
        - RDG - Reuniones
      parameters:
        - name: id 
          in: path
          description: ID de la Reunion
          type: string
          required: true 
        - in: body
          name: body
          description: Datos de la serie de reunion de RDG
          schema:
            type: object
            required:
              - reunion
              - usuarioCreacion
              - fechaCreacion
            properties:
              reunion:
                type: string
              fecha:
                type: string
              desdeHora:
                type: string
              hastaHora:
                type: string
              desdeDate:
                type: integer
              hastaDate:
                type: integer
              usuarioCreacion:
                type: string
              fechaCreacion:
                type: string
              usuarioModificacion:
                type: string
              fechaModificacion:
                type: string
              ubicacion:
                type: object
                properties:
                  nombre:
                    type: string
              version:
                type: string
      responses:
        200:
          description: ok
  
   # //////////////////////////////////////////////////////////////////// RDG - Temarios
...
  /rdg/temarios/{id}:
    x-swagger-router-controller: rdg
    get:
      summary: Trae Temario de una reunion de RDG por ID
      operationId: getTemarioPorId
      parameters:
        - name: id 
          in: path
          description: ID de la Reunion
          type: string
          required: true 
      tags:
        - RDG - Temarios
      responses: 
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Temario'
    put:
      summary: Editar una reunion de RDG por ID
      operationId: updateTemario
      tags:
        - RDG - Temarios
      parameters:
        - name: id 
          in: path
          description: ID de la Reunion
          type: string
          required: true 
        - in: body
          name: body
          description: Datos de la serie de reunion de RDG
          schema:
            type: object
            required:
              - instancia
              - fecha
              - usuario
            properties:
              instancia:
                type: string
              tipoTemario: 
                type: string
              html:
                type: string
              fecha: 
                type: string
              usuario: 
                type: string
              comentarios: 
                type: string
              fechaComentarios: 
                type: string
              propuestaEnviado:
                type: object
                properties:
                  fecha: 
                    type: string
                  version: 
                    type: string
                  para:
                    type: array
                    items:
                      type: object
                  cc:
                    type: array
                    items:
                      type: object
                  cco:
                    type: array
                    items:
                      type: object
                  exclusivos:
                    type: array
                    items:
                      type: object
                  asunto: 
                    type: string
                  mensajeHtml: 
                    type: string
                  usuario: 
                    type: string
              enviado:
                type: object
                properties:
                  fecha: 
                    type: string
                  version: 
                    type: string
                  para:
                    type: array
                    items:
                      type: object
                  cc:
                    type: array
                    items:
                      type: object
                  cco:
                    type: array
                    items:
                      type: object
                  exclusivos:
                    type: array
                    items:
                      type: object
                  asunto: 
                    type: string
                  mensajeHtml: 
                    type: string
                  usuario: 
                    type: string
      responses:
        200:
          description: ok

  # ////////////////////////////////////////////////////////////////////
...
definitions:
...
  Cita:
    type: object
    properties:
      id:
        type: string
      para:
        type: string
      cc:
        type: string
      cco:
        type: string
      usuario:
        type: string
      asunto:
        type: string
      mensajeHtml:
        type: string
      idInstancia:
        type: string
      version: 
        type: string
      fecha:
        type: string
  Minuta:
    type: object
    properties:
      instancia:
        type: string
      compromisos:
        type: array
        items:
          type: object
      html:
        type: ['string','null']
      usuario:
        type: string
      fecha:
        type: string
      tipoMinuta:
        type: ['string','null']
      presentacion:
        type: ['boolean','null']
      enviado:
        type: ['array','null']
        items:
          type: object
  SerieReunion:
    type: object
    properties:
      nombre: 
        type: string
      tipo:
        type: string
      frecuencia:
        type: string
      participantes:
        type: array
        items:
          type: object
      cita:
        type: array
        items:
          type: object
          properties:
            para:
              type: array
              items:
                type: object
            cc:
              type: array
              items:
                type: object
            cco:
              type: array
              items:
                type: object
            exclusivos:
              type: array
              items:
                type: object
      temario:
        type: array
        items:
          type: object
          properties:
            para:
              type: array
              items:
                type: object
            cc:
              type: array
              items:
                type: object
            cco:
              type: array
              items:
                type: object
            exclusivos:
              type: array
              items:
                type: object
      minuta:
        type: array
        items:
          type: object
          properties:
            para:
              type: array
              items:
                type: object
            cc:
              type: array
              items:
                type: object
            cco:
              type: array
              items:
                type: object
            exclusivos:
              type: array
              items:
                type: object
      propuestaTemario:
        type: array
        items:
          type: object
          properties:
            para:
              type: array
              items:
                type: object
            cc:
              type: array
              items:
                type: object
            cco:
              type: array
              items:
                type: object
            exclusivos:
              type: array
              items:
                type: object
      jurisdiccion:
        type: string
      asistenteTablero:
        type: string
      validaTemario:
        type: string
      jurisdicciones:
        type: array
        items:
          type: string
...
  Reunion:
    type: object
    properties:
      reunion:
        type: string
      fecha:
        type: string
      desdeHora:
        type: string
      hastaHora:
        type: string
      desdeDate:
        type: integer
      hastaDate:
        type: integer
      usuarioCreacion:
        type: string
      fechaCreacion:
        type: string
  Temario:
    type: object
    properties:
      instancia:
        type: string
      tipoTemario: 
        type: string
      html:
        type: string
      fecha: 
        type: string
      usuario: 
        type: string
      comentarios: 
        type: string
      fechaComentarios: 
        type: string
      propuestaEnviado:
        type: object
        properties:
          fecha: 
            type: string
          version: 
            type: string
          para:
            type: array
            items:
              type: object
          cc:
            type: array
            items:
              type: object
          cco:
            type: array
            items:
              type: object
          exclusivos:
            type: array
            items:
              type: object
          asunto: 
            type: string
          mensajeHtml: 
            type: string
          usuario: 
            type: string
      enviado:
        type: object
        properties:
          fecha: 
            type: string
          version: 
            type: string
          para:
            type: array
            items:
              type: object
          cc:
            type: array
            items:
              type: object
          cco:
            type: array
            items:
              type: object
          exclusivos:
            type: array
            items:
              type: object
          asunto: 
            type: string
          mensajeHtml: 
            type: string
          usuario: 
            type: string
      
          
        
# To apply Basic auth to the whole API: